New messaging system!!!
The messaging system has changed (big time). Every component now has two HandleMessage methods. One, that is virtual, to override and implement code to handle messages. The message in this method is passed along as an object of type Message. The second method is to call HandleMessage on a component.